consume via npm
