--- bower.json
@@ -1,27 +1,0 @@
-{
-  "name": "knockout-magic-textarea",
-  "version": "1.0.2",
-  "description": "Simple knockout binding for magical auto-sizing textarea heights.",
-  "main": "knockout.magic.textarea.js",
-  "keywords": [
-    "knockout",
-    "textarea",
-    "automatic resize"
-  ],
-  "authors": [
-    "Archon Information Systems"
-  ],
-  "license": "MIT",
-  "homepage": "https://github.com/ArchonInfoSys/knockout-magic-textarea",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ],
-  "dependencies": {
-    "knockout": "~3.1.0",
-    "jquery": "~1.10"
-  }
-}
